+++ base/strings/search.jl
@@ -535,7 +535,7 @@ Find the last occurrence of `pattern` in `string`. Equivalent to
 julia> findlast("o", "Hello to the world")
 15:15
 
-julia> findfirst("Julia", "JuliaLang")
+julia> findlast("Julia", "JuliaLang")
 1:5
 ```
 """